Use internal kinematics of motion gen, remove robot representation
